renaming + refactor
