+++ libs/components/src/shared/patterns/index.ts
@@ -1,4 +1,3 @@
 export * from './affix';
 export * from './focus';
 export * from './form-elements';
-export * from './force-error';
